backend/pyproject.toml
[project]
name = "drainage-system-backend"
version = "0.1.0"
description = "Backend for the drainage system stage 1 foundation"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.12",
  "uvicorn[standard]>=0.34.0",
  "sqlalchemy>=2.0.40",
  "psycopg[binary]>=3.2.6",
  "pydantic-settings>=2.8.1",
  "geoalchemy2>=0.18.4",
  "openpyxl>=3.1.5",
  "python-multipart>=0.0.20"
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
